major FPS changes and minor audio updates
The render and update methods now run on their own timers, so the update method is limited to 60 FPS while the render method can run at infinite FPS, creating smoother graphics. Also, minor changes to two audio tracks and a bug was corrected that stopped the background music when the tab was inactive.